Fix migration-related problems in PinChangedEvent and clean up
Refactor event handling to use generic base class first developed for Event_Generic
